+++ src/ska_dlm_client/register_storage_location/__init__.py
@@ -10,16 +10,9 @@ __author__ = """Mark Boulton"""
 __email__ = "mark.boulton@uwa.edu.au"
 __version__ = "1.0.0"
 
-from ska_dlm_client.register_storage_location.main import (
-    get_or_init_location,
-    get_or_init_storage,
-    main,
-    setup_testing,
-)
+from ska_dlm_client.register_storage_location.main import get_or_init_location, get_or_init_storage
 
 __all__ = [
-    "main",
-    "setup_testing",
     "get_or_init_location",
     "get_or_init_storage",
 ]